Move style parsing from json to styler service
